+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Deprecations
 
 Bug fixes and small changes
 ---------------------------
+- Enable uniform binning for n-dimensional distributions with integer(s).
 
 Experimental
 ------------