+++ CHANGELOG.rst
@@ -21,6 +21,8 @@ Bug fixes and small changes
 ---------------------------
 - Enable uniform binning for n-dimensional distributions with integer(s).
 - Sum of histograms failed for calling the pdf method (can be indirectly), integrated over wrong axis.
+- Binned PDFs expected binned spaces for limits, now unbinned limits are also allowed and automatically
+    converted to binned limits using the PDFs binning.
 
 Experimental
 ------------